create config.js
